feat: dmx out rendered thread added

+++ daemon/src/main.c
@@ -2,14 +2,16 @@
 #include "stage.h"
 #include "clock.h"
 #include "dmx/dmx.h"
+#include "dmx/dmx_out.h"
 
+#include <stdbool.h>
 #include <string.h>
 #include <stdio.h>
 #include <signal.h>
 
 #define SPARKD_VERSION "0.0.0"
 
-#define PERIOD_MS (uint32_t)(1000 / 40)
+#define PERIOD_MS (uint32_t)(50)
 
 static volatile uint8_t should_keep_running = 1;
 
@@ -21,18 +23,18 @@ void signal_handler(int signum)
 
 typedef struct {
     spark_log_level_t log_level;
-    uint8_t print_help;
-    uint8_t print_version;
+    bool print_help;
+    bool print_version;
 } spark_args_t;
 
-void parse_cmdline_args(int argc, char **argv, spark_args_t *args)
+static void parse_cmdline_args(int argc, char **argv, spark_args_t *args)
 {
     for (int i=0 ; i<argc ; i++)
     {
         if (strcmp("--help", argv[i]) == 0)
-            args->print_help = 1;
+            args->print_help = true;
         else if (strcmp("--version", argv[i]) == 0)
-            args->print_version = 1;
+            args->print_version = true;
         else if (strcmp("--log-level", argv[i]) == 0 && i + 1 < argc)
         {
             spark_log_level_from_string(argv[i + 1], &args->log_level);
@@ -44,8 +46,8 @@ void parse_cmdline_args(int argc, char **argv, spark_args_t *args)
 int main(int argc, char **argv)
 {
     spark_args_t args = {
-        .print_help = 0,
-        .print_version = 0,
+        .print_help = false,
+        .print_version = false,
         .log_level = SPARK_LOG_INFO
     };
     parse_cmdline_args(argc, argv, &args);
@@ -68,9 +70,9 @@ int main(int argc, char **argv)
         return 0;
     }
 
-    spark_log_init(args.log_level);
+    int rc;
 
-    uint8_t frame[SPARK_DMX_UNIVERSE_SIZE];
+    spark_log_init(args.log_level);
 
     spark_stage_t stage;
     spark_stage_init(&stage);
@@ -80,8 +82,15 @@ int main(int argc, char **argv)
     spark_dmx_dummy_init(&dmx_backend);
     spark_log_debug("Dummy DMX backend initialized!");
 
-    dmx_backend.ops->open(&dmx_backend);
-    spark_log_debug("DMX backend opened!");
+    spark_dmx_out_t dmx_out;
+    spark_dmx_out_init(&dmx_out, &dmx_backend, &stage);
+    rc = spark_dmx_out_start(&dmx_out);
+    if (rc != 0)
+    {
+        spark_log_error("DMX output thread failed to start (%d)", rc);
+        return rc;
+    }
+    spark_log_debug("DMX thread started!");
     
     signal(SIGINT, signal_handler);   /* Ctrl+C */
     signal(SIGTERM, signal_handler);  /* kill command */
@@ -89,20 +98,18 @@ int main(int argc, char **argv)
     spark_log_info("Ctrl+C to stop...");
     while (should_keep_running)
     {
-        uint64_t start = spark_clock_monotonic_ms();
-
-        spark_stage_render(&stage, frame);
-
-        dmx_backend.ops->send_frame(&dmx_backend, frame);
-
-        uint32_t elapsed = spark_clock_monotonic_ms() - start;
-        if (elapsed < PERIOD_MS)
-            spark_clock_msleep(PERIOD_MS - (uint32_t)elapsed);
+        spark_clock_msleep(PERIOD_MS);
     }
 
     spark_log_info("Shutting down");
 
-    dmx_backend.ops->close(&dmx_backend);
+    rc = spark_dmx_out_stop(&dmx_out);
+    if (rc != 0)
+        spark_log_error("DMX out thread stop error (%d)", rc);
+    else
+        spark_log_debug("DMX out thread stopped!");
+
+    spark_dmx_close(&dmx_backend);
     spark_log_debug("DMX backend closed!");
 
     spark_stage_destroy(&stage);
